Movement branch creation and updates
Created movement branch to work on the movement of the buggy. Made a set of functions for set pieces of operation, which can be tuned separately. The moveColor functions then use these set pieces as inputs so that the tuning can all be done centrally. Lastly, made a testing framework in the main file so that the amount of turning and directional movement can be tuned for each movement function. Additionally, the car would only move when RF3 is pressed.